Combined process tomography tests and added some additional checks
